# backend/api/openapi.yaml
openapi: "3.1.0"
info:
  title: SpendWise API
  description: Collaborative household financial ledger API
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/workspaces/{workspaceId}/rules:
    post:
      operationId: createRule
      tags: [Rules]
      summary: Create a categorization rule (editor+ required)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRuleRequest"
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    get:
      operationId: listRules
      tags: [Rules]
      summary: List categorization rules (viewer+ required)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: List of rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RuleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/workspaces/{workspaceId}/rules/{ruleId}:
    put:
      operationId: updateRule
      tags: [Rules]
      summary: Update a categorization rule (editor+ required)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/RuleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRuleRequest"
      responses:
        "200":
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteRule
      tags: [Rules]
      summary: Delete a categorization rule (editor+ required)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/RuleId"
      responses:
        "204":
          description: Rule deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────
  # Summary
  # ──────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from `/api/v1/auth/{provider}/callback`
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    RuleId:
      name: ruleId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "workspace not found"
      required: [error]
...
    CreateRuleRequest:
      type: object
      properties:
        match_pattern:
          type: string
          description: Text pattern to match against transaction description
          example: "biedronka"
        target_category_id:
          type: integer
          format: int64
        priority:
          type: integer
          default: 0
          description: Higher priority rules take precedence
      required: [match_pattern, target_category_id]
    UpdateRuleRequest:
      type: object
      properties:
        match_pattern:
          type: string
        target_category_id:
          type: integer
          format: int64
        priority:
          type: integer
      required: [match_pattern, target_category_id]
    RuleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        workspace_id:
          type: integer
          format: int64
        match_pattern:
          type: string
        target_category_id:
          type: integer
          format: int64
        priority:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, workspace_id, match_pattern, target_category_id, priority, created_at, updated_at]

    # ── Summary ──
...